fix: [sc-17] move middleware from auth to separate files
