+++ src/routes/expenses.route.js
@@ -14,20 +14,19 @@ const {
   deleteComment,
 } = require('../controllers/expenses.controller.js');
 
-const {
-  verifyToken,
-  checkUserInGroup,
-} = require('../middlewares/auth.middleware.js');
+const { verifyToken } = require('../middlewares/auth.middleware.js');
+
+const { checkUserInGroup } = require('../middlewares/expenses.middleware.js');
 
 router.post('/', verifyToken, checkUserInGroup, createExpense);
 
 router.get('/', verifyToken, checkUserInGroup, getAllExpenses);
 
-router.get('/:expenseId', verifyToken, checkUserInGroup, getExpenseDetails);
+router.get('/:expenseId', verifyToken, getExpenseDetails);
 
-router.put('/:expenseId', verifyToken, checkUserInGroup, updateExpense);
+router.put('/:expenseId', verifyToken, updateExpense);
 
-router.delete('/:expenseId', verifyToken, checkUserInGroup, deleteExpense);
+router.delete('/:expenseId', verifyToken, deleteExpense);
 
 router.post('/:expenseId/settle-up', verifyToken, settleUpExpense);
